requirements.txt
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
apricot-select==0.6.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
comm==0.2.3
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
hf-xet==1.3.2
httplib2==0.31.2
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lightning-utilities==0.15.2
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mmcv==1.5.2
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.2.1
nose==1.3.7
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.54.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wandb==0.25.0
wcwidth==0.6.0
yapf==0.40.1
yarl==1.22.0
zipp==3.23.0
